Add profile menu transition
